stack Menu 90%
the stack menu now allows the user to highlight up to four peices in the
stack whil displaying 5, the first peice is always highlighted.

I now need to implemnt the black and white piece difference, as well as
account for when stacks are less than 5 deep. the select button will
need to return the value of the amount of pieces selected but that will
just be an int. Back button works and removes all values but should also
reset the values.